move active-profile-view and edit profile view
